Adding the review report
This GitHub repository contains a comprehensive report on the latest advancements in smart wheelchairs, authored by Ahmad Rahmani. Covering papers from 2016 to 20121, this review provides a concise overview of the current state of hardware and software technologies, presents key findings and results, and highlights the significance of independent mobility for individuals relying on wheelchairs.